fix: 🐛 convert group availability to boolean representation
