+++ src/lib/components/availability/PersonalAvailability.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-  import { onMount } from "svelte";
-
   import type { PageData } from "../../../routes/availability/$types";
 
   import LoginFlow from "./LoginModal.svelte";
@@ -9,13 +7,11 @@
   import {
     availabilityDates,
     availabilityTimeBlocks,
-    guestSession,
     isEditingAvailability,
     isStateUnsaved,
   } from "$lib/stores/availabilityStores";
   import type { AvailabilityBlockType, SelectionStateType } from "$lib/types/availability";
   import { ZotDate } from "$lib/utils/ZotDate";
-  import { getGeneralAvailability } from "$lib/utils/availability";
   import { cn } from "$lib/utils/utils";
 
   export let columns: number;
@@ -152,19 +148,6 @@
       }
     });
   }
-
-  onMount(async () => {
-    $guestSession.meetingId = data.meetingId;
-
-    const generalAvailability = await getGeneralAvailability(data, $guestSession);
-    const defaultMeetingDates = data.defaultDates.map((item) => new ZotDate(item.date, false, []));
-    ZotDate.initializeAvailabilities(defaultMeetingDates);
-
-    $availabilityDates =
-      generalAvailability && generalAvailability.length > 0
-        ? generalAvailability
-        : defaultMeetingDates;
-  });
 </script>
 
 <div class="flex items-center justify-between overflow-x-auto font-dm-sans">
